reintroduced long long

+++ Avisynth3/trunk/avisynth/src/core/bufferwindow/blender.h
@@ -52,9 +52,9 @@ template <int bps> class Blender { };
 template <> class Blender<1>
 {
 
-  int64 weight64_;
-  bool noBlend_;       //true if the blend is a no op
-  bool fullBlend_;     //true if the blend is a copy from blendFrom to blendIn
+  long long weight64_;
+  bool noBlend_;         //true if the blend is a no op
+  bool fullBlend_;       //true if the blend is a copy from blendFrom to blendIn
 
 
 public:  //structors
